+++ src/types.ts
@@ -12,23 +12,23 @@ export interface ModuleOptions {
    */
   classDefault?: string
   /**
-   * Clear all classes from SVGs.
+   * Clear all classes from SVGs (can be just all, just parent or none).
    *
-   * @default false
+   * @default 'none'
    */
-  clearClasses: boolean
+  clearClasses: 'all' | 'parent' | 'none'
   /**
-   * Clear native `width` and `height` attributes from SVGs.
+   * Clear native `width` and `height` attributes from SVGs (can be just all, just parent or none).
    *
-   * @default true
+   * @default 'none'
    */
-  clearSize: boolean
+  clearSize: 'all' | 'parent' | 'none'
   /**
-   * Clear all native styles from SVGs.
+   * Clear all native styles from SVGs (can be just all, just parent or none).
    *
-   * @default false
+   * @default 'none'
    */
-  clearStyles: boolean
+  clearStyles: 'all' | 'parent' | 'none'
   /**
    * Name of the component to use in your application
    *
@@ -47,6 +47,12 @@ export interface ModuleOptions {
    * @default '<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="red" style="width: 1.5rem; height: 1.5rem;"><path stroke-linecap="round" stroke-linejoin="round" d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126zM12 15.75h.007v.008H12v-.008z" /></svg>'
    */
   fallback: string | false
+  /**
+   * Add `display: inline-block` to component.
+   *
+   * @default true
+   */
+  inlineBlock: boolean
   /**
    * Global option to toggle lazy icons, if disabled, hot reloading can have some issues.
    *